Fixed some layout issues with the portfolio section where pictures would cover the project descriptions on small screen laptops. The fix was a bit forced (just a mere responsive font size change) so I'll have to look into making pictures maintain aspect ratio regardless of screen size.
Projects now properly link to the github repos, also the readme's of each repo
now include a link to a live version of the project. Instructions are also
provided to checking out projects that were not hosted on Github.

Added a media query for large laptops to have larger font size in project description.

Just need to add a personal photograph.